Dockerfile cleanup (#1352)
* Dockerfile.multiarch: use bind- and cache-mounts to speedup build

Use a cache mount for go build cache in the build container and mount
the sources as a bind mount as recommended in
https://docs.docker.com/build/cache/optimize/ this largely speeds up the
container build for subsequent builds.

* Use Dockerfile.multiarch for "dev-docker" target

Let's remove some redundancy. AFAICS the Docker.multiarch does
everything the Docker.linux.* files did. And with the build caches
enable it should be just as quick as building on the host.

* Dockerfile.multiarch: Align the alpine version of the base images

* Dockerfile: Reduce build context by adding more files to .dockerignore